Revert "CIV-16786 Added LipvLR tests and enable GALip tests in Prod (#5853)"
This reverts commit c570ddc1330759d24a3300cb25ddb14971eb2a0b.
