+++ src/test/functionalTests/tests/nonprod/LipVLR_e2e_tests.js
@@ -92,53 +92,3 @@ Scenario('LipVLR - NoC and DefendantLR respond as AdmitAll @citizenUI - @api @no
     await api.defendantLRResponse(config.defendantSolicitorUser, 'FULL_ADMISSION', camundaEvent, expectedState);
   }
 });
-
-Scenario('LipVLR - NoC and DefendantLR respond as PartAdmit @citizenUI - @api @noc @nightly', async ({
-  I,
-  api,
-}) => {
-  if (['preview', 'demo'].includes(config.runningEnv)) {
-    await nocSteps.requestNoticeOfChangeForRespondent1Solicitor(claimRef, defendantName, config.defendantSolicitorUser);
-    await api.checkUserCaseAccess(config.defendantCitizenUser, false);
-    await api.checkUserCaseAccess(config.defendantSolicitorUser, true);
-
-    await LoginSteps.EnterCitizenCredentials(config.claimantCitizenUser.email, config.claimantCitizenUser.password);
-    await I.amOnPage('/dashboard');
-    await I.click(legacyCaseReference);
-
-    if (isDashboardServiceEnabled) {
-      const notification = nocForLip(defendantName);
-      await verifyNotificationTitleAndContent(legacyCaseReference, notification.title, notification.content);
-      await I.click(notification.nextSteps);
-    }
-
-    camundaEvent = 'APPLY_NOC_DECISION_DEFENDANT_LIP';
-    expectedState = 'PROCEEDS_IN_HERITAGE_SYSTEM';
-    await api.defendantLRResponse(config.defendantSolicitorUser, 'PART_ADMISSION', camundaEvent, expectedState);
-  }
-});
-
-Scenario('LipVLR - NoC and DefendantLR respond as CounterClaim @citizenUI - @api @noc @nightly', async ({
-  I,
-  api,
-}) => {
-  if (['preview', 'demo'].includes(config.runningEnv)) {
-    await nocSteps.requestNoticeOfChangeForRespondent1Solicitor(claimRef, defendantName, config.defendantSolicitorUser);
-    await api.checkUserCaseAccess(config.defendantCitizenUser, false);
-    await api.checkUserCaseAccess(config.defendantSolicitorUser, true);
-
-    await LoginSteps.EnterCitizenCredentials(config.claimantCitizenUser.email, config.claimantCitizenUser.password);
-    await I.amOnPage('/dashboard');
-    await I.click(legacyCaseReference);
-
-    if (isDashboardServiceEnabled) {
-      const notification = nocForLip(defendantName);
-      await verifyNotificationTitleAndContent(legacyCaseReference, notification.title, notification.content);
-      await I.click(notification.nextSteps);
-    }
-
-    camundaEvent = 'APPLY_NOC_DECISION_DEFENDANT_LIP';
-    expectedState = 'PROCEEDS_IN_HERITAGE_SYSTEM';
-    await api.defendantLRResponse(config.defendantSolicitorUser, 'COUNTER_CLAIM', camundaEvent, expectedState);
-  }
-});
